Create a multi-line, printable console message from an object.

Function takes glue input and properly collpases it using "" to make it multi-line string. Using message without properly collapsing will create one long, flat message string. Nice when printing small tables.

Usage

glue_collapse(msg)

Arguments

msg

a string using str_glue syntax to print a df like object

Value

a string congaing "" to force string to print on multiple lines

Examples

temp = data.frame(
highway = c("90 East", "405 Northbound", "101")
,count = c(100, 205, 92)
) %>%
 mutate(percent = round(count/sum(count), 2))

str_glue("{temp$count} ({temp$percent}%) - {temp$highway}") %>%
 message()
#> 100 (0.25%) - 90 East205 (0.52%) - 405 Northbound92 (0.23%) - 101

glue_collapse("{temp$count} ({temp$percent}%) - {temp$highway}") %>%
 message()
#> Error: Failed to evaluate glue component {temp$count}
#> Caused by error:
#> ! object 'temp' not found