make_diff

Makes difference column from a dataframe column defined by the user using lags as defined by the user. Should not be used directly by user - it is better to use via auto_make_diff_col() function.

Usage

make_diff(col, lags, percent = F)

Arguments

col

a column to make lags

lags

numeric list

percent

boolean to determine if difference should be percent difference

Value

dataframe with differenced column

Examples

data.frame(index = 1:10,
          value1 = rnorm(10), value2 = rnorm(10)) %>%
 mutate(value1_differenced = make_diff(col = index, c(1), percent = F)) %>%
 mutate(value1_differenced_percent = make_diff(col = index, c(1), percent = T))
#>    index      value1     value2 value1_differenced value1_differenced_percent
#> 1      1  0.23669628  0.5562243                 NA                         NA
#> 2      2  1.31829338 -0.5482573                  1                  1.0000000
#> 3      3  0.52390979  1.1105349                  1                  0.5000000
#> 4      4  0.60674805 -2.6123343                  1                  0.3333333
#> 5      5 -0.10993567 -0.1556938                  1                  0.2500000
#> 6      6  0.17218172  0.4338898                  1                  0.2000000
#> 7      7 -0.09032729 -0.3819511                  1                  0.1666667
#> 8      8  1.92434334  0.4241876                  1                  0.1428571
#> 9      9  1.29839276  1.0631020                  1                  0.1250000
#> 10    10  0.74879127  1.0487126                  1                  0.1111111