Create a Shiny button with popover message

This function generates a Shiny button along with a popover message. The button, when pressed or hovered over (based on the specified trigger), displays a popover message with the provided title and text. The default behavior is to show an info-style message with a question mark icon.

Usage

shiny_bs_popover(id, title, text, trigger = "hover")

Arguments

id

The unique identifier for the button and popover.

title

The title of the popover message.

text

The text content of the popover message.

trigger

The trigger for displaying the popover (default is "hover").

Value

A Shiny tagList containing the button and popover components.

See also

Examples

if (FALSE) { # \dontrun{
  shiny_bs_popover(id = "example_button", title = "Information",
                   text = "This is an example popover message.")
} # }