Order factor levels numerically

`strg_numeric_order` orders factor levels numerically based on the values in the input vector. If `rev` is set to `TRUE`, the levels are reversed.

Usage

strg_numeric_order(col, rev = TRUE)

Arguments

col

A factor or character vector.

rev

Logical; if `TRUE`, reverse the order of levels.

Value

A factor with reordered levels.

Details

This function is useful for ensuring that factor levels are ordered correctly when dealing with numeric values. It leverages the `stringr::str_sort` function to sort the levels.

Examples

temp_df = paste0("lab_", seq(1, 50, 2)) %>%
data.frame(label_unordered = .
           ,label_ordered = strg_numeric_order(.))

temp_df %>% arrange(label_unordered) %>% print()
#>    label_unordered label_ordered
#> 1            lab_1         lab_1
#> 2           lab_11        lab_11
#> 3           lab_13        lab_13
#> 4           lab_15        lab_15
#> 5           lab_17        lab_17
#> 6           lab_19        lab_19
#> 7           lab_21        lab_21
#> 8           lab_23        lab_23
#> 9           lab_25        lab_25
#> 10          lab_27        lab_27
#> 11          lab_29        lab_29
#> 12           lab_3         lab_3
#> 13          lab_31        lab_31
#> 14          lab_33        lab_33
#> 15          lab_35        lab_35
#> 16          lab_37        lab_37
#> 17          lab_39        lab_39
#> 18          lab_41        lab_41
#> 19          lab_43        lab_43
#> 20          lab_45        lab_45
#> 21          lab_47        lab_47
#> 22          lab_49        lab_49
#> 23           lab_5         lab_5
#> 24           lab_7         lab_7
#> 25           lab_9         lab_9
temp_df %>% arrange(desc(label_ordered)) %>% print()
#>    label_unordered label_ordered
#> 1            lab_1         lab_1
#> 2            lab_3         lab_3
#> 3            lab_5         lab_5
#> 4            lab_7         lab_7
#> 5            lab_9         lab_9
#> 6           lab_11        lab_11
#> 7           lab_13        lab_13
#> 8           lab_15        lab_15
#> 9           lab_17        lab_17
#> 10          lab_19        lab_19
#> 11          lab_21        lab_21
#> 12          lab_23        lab_23
#> 13          lab_25        lab_25
#> 14          lab_27        lab_27
#> 15          lab_29        lab_29
#> 16          lab_31        lab_31
#> 17          lab_33        lab_33
#> 18          lab_35        lab_35
#> 19          lab_37        lab_37
#> 20          lab_39        lab_39
#> 21          lab_41        lab_41
#> 22          lab_43        lab_43
#> 23          lab_45        lab_45
#> 24          lab_47        lab_47
#> 25          lab_49        lab_49